chore(bulk-import): optimize bulk-import
