+++ workspaces/bulk-import/.changeset/smart-crews-wish.md
@@ -1,0 +1,5 @@
+---
+'@red-hat-developer-hub/backstage-plugin-bulk-import': patch
+---
+
+optimize bulk-import
