+++ workspaces/bulk-import/plugins/bulk-import/src/utils/repository-utils.tsx
@@ -19,16 +19,15 @@ import { Link, StatusOK } from '@backstage/core-components';
 import { configApiRef, useApi } from '@backstage/core-plugin-api';
 
 import Typography from '@mui/material/Typography';
-import * as jsyaml from 'js-yaml';
+import { loadAll } from 'js-yaml';
 import { get } from 'lodash';
-import * as yaml from 'yaml';
+import { stringify } from 'yaml';
 import * as yup from 'yup';
 
 import { WaitingForPR } from '../components/WaitingForPR';
 import {
   AddedRepositories,
   AddRepositoryData,
-  APITypes,
   ApprovalTool,
   CreateImportJobRepository,
   ErrorType,
@@ -49,6 +48,8 @@ import {
 } from '../types';
 import { getWorkflowStatusInfo } from './orchestratorStatus';
 import { getTaskStatusInfo } from './task-status';
+
+export { getApi } from './getApi';
 
 export const TaskLink = ({
   taskId,
@@ -521,7 +522,7 @@ export const prepareDataForSubmission = (
           organization: repo.orgName || '',
           defaultBranch: repo.defaultBranch || '',
         },
-        catalogInfoContent: yaml.stringify(
+        catalogInfoContent: stringify(
           repo.catalogInfoYaml?.prTemplate?.yaml,
           null,
           2,
@@ -539,31 +540,6 @@ export const prepareDataForSubmission = (
     },
     [],
   );
-
-export const getApi = (
-  backendUrl: string,
-  page: number,
-  size: number,
-  searchString: string,
-  approvalTool: string,
-  options?: APITypes,
-) => {
-  const params = new URLSearchParams({
-    pagePerIntegration: String(page),
-    sizePerIntegration: String(size),
-    search: searchString,
-    approvalTool,
-  });
-
-  if (options?.fetchOrganizations) {
-    return `${backendUrl}/api/bulk-import/organizations?${params.toString()}`;
-  }
-  if (options?.orgName) {
-    const orgName = encodeURIComponent(options?.orgName);
-    return `${backendUrl}/api/bulk-import/organizations/${orgName}/repositories?${params.toString()}`;
-  }
-  return `${backendUrl}/api/bulk-import/repositories?${params.toString()}`;
-};
 
 export const getCustomisedErrorMessage = (
   status: (RepositoryStatus | string)[] | undefined,
@@ -636,7 +612,7 @@ export const evaluatePRTemplate = (
 ): { pullReqPreview: PullRequestPreview; isInvalidEntity: boolean } => {
   const gitProvider = isGithubJob(repositoryStatus) ? 'github' : 'gitlab';
   try {
-    const entity = jsyaml.loadAll(
+    const entity = loadAll(
       repositoryStatus[gitProvider]?.pullRequest.catalogInfoContent ?? '',
     )[0] as Entity;
     const isInvalid =
